Diagram visual refactoring
