+++ app/components/process/diagram/visual.hbs
@@ -1,24 +1,20 @@
-{{! Latest diagram }}
-
-{{! Latest diagram - Title }}
 <div class="au-u-flex au-u-flex--between au-u-margin-bottom-small">
   <AuHeading @level="2" @skin="3">Diagram</AuHeading>
   <AuButtonGroup>
     {{#if @canEdit}}
       <AuButton
         @skin="secondary"
-        {{on "click" @openReplaceModal}}
+        {{on "click" this.openReplaceModal}}
       >Vervang</AuButton>
     {{/if}}
-    {{#if @latestDiagram}}
-      <AuButton {{on "click" @openDownloadModal}}>
+    {{#if this.latestDiagram}}
+      <AuButton {{on "click" this.openDownloadModal}}>
         Download
       </AuButton>
     {{/if}}
   </AuButtonGroup>
 </div>
 
-{{! Latest diagram - Visual }}
 <div>
   {{#if this.latestDiagramIsLoading}}
     <AuLoader />
@@ -28,21 +24,22 @@
     </AuAlert>
   {{else if
     (and
-      @latestDiagram (or @latestDiagram.isBpmnFile @latestDiagram.isVisioFile)
+      this.latestDiagram
+      (or this.latestDiagram.isBpmnFile this.latestDiagram.isVisioFile)
     )
   }}
     <AuHelpText @skin="tertiary" class="au-u-margin-top-none">
       Klik en hou vast om diagram te bedienen. Gebruik ctrl en scrollwiel om in
       en uit te zoomen. Klik elders om bediening uit te schakelen.
     </AuHelpText>
-    {{#if @latestDiagram.isBpmnFile}}
+    {{#if this.latestDiagram.isBpmnFile}}
       <BpmnViewer
-        @diagram={{@latestDiagram}}
-        @onBpmnLoaded={{@setLatestDiagramAsBpmn}}
-        @onSvgLoaded={{@setLatestDiagramAsSvg}}
+        @diagram={{this.latestDiagram}}
+        @onBpmnLoaded={{this.setLatestDiagramAsBpmn}}
+        @onSvgLoaded={{this.setLatestDiagramAsSvg}}
       />
-    {{else if @latestDiagram.isVisioFile}}
-      <PdfViewer @diagram={{@latestDiagram}} />
+    {{else if this.latestDiagram.isVisioFile}}
+      <PdfViewer @diagram={{this.latestDiagram}} />
     {{/if}}
     <DataCard class="au-u-margin-top-small">
       <:header></:header>
@@ -55,13 +52,13 @@
             <Item>
               <:label>Bestandsnaam</:label>
               <:content>
-                {{@latestDiagram.name}}
+                {{this.latestDiagram.name}}
               </:content>
             </Item>
             <Item>
               <:label>Toegevoegd op</:label>
               <:content>
-                {{date-format @latestDiagram.modified true}}
+                {{date-format this.latestDiagram.modified true}}
               </:content>
             </Item>
           </:left>
@@ -69,19 +66,19 @@
             <Item>
               <:label>Grootte</:label>
               <:content>
-                {{file-size-format @latestDiagram.size}}
+                {{file-size-format this.latestDiagram.size}}
               </:content>
             </Item>
             <Item>
               <:label>Extensie</:label>
               <:content>
-                {{@latestDiagram.extension}}
+                {{this.latestDiagram.extension}}
               </:content>
             </Item>
             <Item>
               <:label>Formaat</:label>
               <:content>
-                {{@latestDiagram.format}}
+                {{this.latestDiagram.format}}
               </:content>
             </Item>
           </:right>
@@ -94,3 +91,89 @@
     </AuAlert>
   {{/if}}
 </div>
+
+<AuModal
+  @modalOpen={{this.replaceModalOpened}}
+  @closeModal={{this.closeReplaceModal}}
+>
+  <:title>Vervang diagram</:title>
+  <:body>
+    <CustomFileUpload
+      @accept=".bpmn,.vsdx"
+      @title="BPMN- of Visiobestand selecteren"
+      @helpTextDragDrop="Bestand wordt meteen verwerkt en toegevoegd aan bibliotheek na selectie"
+      @updateProcess={{this.addFileToProcess}}
+      @extractBpmnElements={{this.extractBpmnElements}}
+      @onFinishUpload={{this.diagramUploaded}}
+    />
+  </:body>
+  <:footer>
+    <AuButtonGroup class="au-u-flex--end">
+      <AuButton
+        @skin="naked"
+        {{on "click" this.closeReplaceModal}}
+      >Annuleer</AuButton>
+    </AuButtonGroup>
+  </:footer>
+</AuModal>
+
+<AuModal
+  @modalOpen={{this.downloadModalOpened}}
+  @closeModal={{this.closeDownloadModal}}
+>
+  <:title>Download diagram</:title>
+  <:body>
+    <div class="au-o-flow">
+      <p>Het geselecteerde diagram is beschikbaar in verschillende formaten:</p>
+      <AuButtonGroup class="au-u-flex au-u-flex--around">
+        {{#if this.latestDiagram.isVisioFile}}
+          <AuButton
+            @skin="primary"
+            {{on "click" (perform this.downloadLatestDiagram "vsdx")}}
+          >
+            Visio
+            <span class="au-u-para-tiny au-u-regular">(.vsdx)</span>
+          </AuButton>
+        {{/if}}
+        <AuButton
+          @skin={{if this.latestDiagram.isBpmnFile "primary" "secondary"}}
+          {{on "click" (perform this.downloadLatestDiagram "bpmn")}}
+        >
+          BPMN
+          <span class="au-u-para-tiny au-u-regular">(.bpmn)</span>
+        </AuButton>
+        {{#if this.latestDiagramAsSvg}}
+          <AuButton
+            @skin="secondary"
+            {{on "click" (perform this.downloadLatestDiagram "png")}}
+          >
+            Afbeelding
+            <span class="au-u-para-tiny au-u-regular">(.png)</span>
+          </AuButton>
+          <AuButton
+            @skin="secondary"
+            {{on "click" (perform this.downloadLatestDiagram "svg")}}
+          >
+            Vectorafbeelding
+            <span class="au-u-para-tiny au-u-regular">(.svg)</span>
+          </AuButton>
+        {{/if}}
+        <AuButton
+          @skin="secondary"
+          {{on "click" (perform this.downloadLatestDiagram "pdf")}}
+        >
+          PDF
+          <span class="au-u-para-tiny au-u-regular">(.pdf)</span>
+        </AuButton>
+      </AuButtonGroup>
+    </div>
+  </:body>
+  <:footer>
+    <AuButtonGroup class="au-u-flex--end">
+      <AuButton
+        @skin="naked"
+        {{on "click" this.closeDownloadModal}}
+      >Annuleer</AuButton>
+    </AuButtonGroup>
+  </:footer>
+</AuModal>
