+++ app/controllers/processes/process/index.js
@@ -1,19 +1,9 @@
 import Controller from '@ember/controller';
 import { tracked } from '@glimmer/tracking';
 import { action } from '@ember/object';
-import { dropTask, enqueueTask, keepLatestTask } from 'ember-concurrency';
+import { enqueueTask, keepLatestTask } from 'ember-concurrency';
 import { inject as service } from '@ember/service';
-import { htmlSafe } from '@ember/template';
-import FileSaver from 'file-saver';
 import ENV from 'frontend-openproceshuis/config/environment';
-import removeFileNameExtension from 'frontend-openproceshuis/utils/file-extension-remover';
-import {
-  convertSvgToPdf,
-  convertSvgToPng,
-} from 'frontend-openproceshuis/utils/svg-convertors';
-import generateFileDownloadUrl, {
-  generateVisioConversionUrl,
-} from 'frontend-openproceshuis/utils/file-download-url';
 
 export default class ProcessesProcessIndexController extends Controller {
   queryParams = [
@@ -42,25 +32,6 @@ export default class ProcessesProcessIndexController extends Controller {
 
   get process() {
     return this.model.process;
-  }
-
-  // Latest diagram
-
-  @tracked latestDiagram = undefined;
-  @tracked latestDiagramIsLoading = true;
-  @tracked latestDiagramHasErrored = false;
-
-  latestDiagramAsBpmn = undefined;
-  latestDiagramAsSvg = undefined;
-
-  @action
-  setLatestDiagramAsBpmn(value) {
-    this.latestDiagramAsBpmn = value;
-  }
-
-  @action
-  setLatestDiagramAsSvg(value) {
-    this.latestDiagramAsSvg = value;
   }
 
   // Diagram versions
@@ -108,86 +79,7 @@ export default class ProcessesProcessIndexController extends Controller {
     this.downloadModalOpened = false;
   }
 
-  @dropTask
-  *downloadLatestDiagram(targetExtension) {
-    try {
-      if (!this.latestDiagram) throw Error;
-
-      let blob = undefined;
-      if (targetExtension === 'vsdx' && this.latestDiagram.isVisioFile) {
-        const url = generateFileDownloadUrl(this.latestDiagram.id);
-        const response = yield fetch(url);
-        if (!response.ok) throw Error(response.status);
-        blob = yield response.blob();
-      } else if (targetExtension === 'bpmn') {
-        if (this.latestDiagramAsBpmn) {
-          blob = new Blob([this.latestDiagramAsBpmn], {
-            type: 'application/xml;charset=utf-8',
-          });
-        } else if (this.latestDiagram.isVisioFile) {
-          const url = generateVisioConversionUrl(
-            this.latestDiagram.id,
-            targetExtension,
-          );
-          const response = yield fetch(url);
-          if (!response.ok) throw Error(response.status);
-          blob = yield response.blob();
-        }
-      } else if (targetExtension === 'svg' && this.latestDiagramAsSvg) {
-        blob = new Blob([this.latestDiagramAsSvg], {
-          type: 'image/svg+xml;charset=utf-8',
-        });
-      } else if (targetExtension === 'pdf') {
-        if (this.latestDiagramAsSvg) {
-          blob = yield convertSvgToPdf(this.latestDiagramAsSvg);
-        } else if (this.latestDiagram.isVisioFile) {
-          const url = generateVisioConversionUrl(
-            this.latestDiagram.id,
-            targetExtension,
-          );
-          const response = yield fetch(url);
-          if (!response.ok) throw Error(response.status);
-          blob = yield response.blob();
-        }
-      } else if (targetExtension === 'png' && this.latestDiagramAsSvg) {
-        blob = yield convertSvgToPng(this.latestDiagramAsSvg);
-      }
-      if (!blob) throw Error;
-
-      const fileName = `${removeFileNameExtension(
-        this.latestDiagram.name,
-        this.latestDiagram.extension,
-      )}.${targetExtension}`;
-
-      FileSaver.saveAs(blob, fileName);
-    } catch {
-      let message = 'Bestand kon niet worden opgehaald';
-
-      if (this.latestDiagram.isVisioFile && targetExtension === 'bpmn') {
-        const mailto =
-          'mailto:loketlokaalbestuur@vlaanderen.be' +
-          `?subject=${encodeURIComponent(
-            'Visio kan niet downloaden als BPMN',
-          )}` +
-          `?body=${encodeURIComponent(`\n\n${window.location.href}\n`)}`;
-        const linkHtml = `<a href="${mailto}">Stuur ons een mailtje</a>`;
-        message = htmlSafe(
-          `Dit visio-bestand kan niet worden gedownload als BPMN. ${linkHtml} met het proces waarover het gaat en een korte beschrijving van wat niet lukt. Dan kunnen we nagaan wat fout gaat en Open Proces Huis verder verbeteren.`,
-        );
-      }
-
-      this.toaster.error(message, 'Fout');
-      return;
-    }
-
-    this.trackDownloadFileEvent(
-      this.latestDiagram.id,
-      this.latestDiagram.name,
-      'bpmn',
-      targetExtension,
-    );
-  }
-
+  @action
   trackDownloadFileEvent(fileId, fileName, fileExtension, targetExtension) {
     this.plausible.trackEvent('Download bestand', {
       'Bestand-ID': fileId,
@@ -211,27 +103,33 @@ export default class ProcessesProcessIndexController extends Controller {
 
   @enqueueTask
   *loadFileDownloads(targetExtension) {
-    const process = yield this.store.findRecord('process', this.process.id);
-    const stats = process.processStatistics;
-
-    switch (targetExtension) {
-      case 'bpmn':
-        stats.bpmnDownloads += 1;
-        break;
-      case 'pdf':
-        stats.pdfDownloads += 1;
-        break;
-      case 'png':
-        stats.pngDownloads += 1;
-        break;
-      case 'svg':
-        stats.svgDownloads += 1;
-        break;
-      default:
-        console.error('fileExtension', targetExtension, 'not recognized');
-        return;
+    try {
+      const process = yield this.store.findRecord('process', this.process.id);
+      const stats = process.processStatistics;
+      switch (targetExtension) {
+        case 'bpmn':
+          stats.bpmnDownloads += 1;
+          break;
+        case 'pdf':
+          stats.pdfDownloads += 1;
+          break;
+        case 'png':
+          stats.pngDownloads += 1;
+          break;
+        case 'svg':
+          stats.svgDownloads += 1;
+          break;
+        default:
+          console.error('fileExtension', targetExtension, 'not recognized');
+          return;
+      }
+      yield stats.save();
+    } catch (error) {
+      console.error(
+        `Something went wrong while loading the file downloads:`,
+        error,
+      );
     }
-    yield stats.save();
   }
 
   @action
@@ -252,22 +150,6 @@ export default class ProcessesProcessIndexController extends Controller {
   @action
   closeAddModal() {
     this.addModalOpened = false;
-  }
-
-  @dropTask
-  *extractBpmnElements(bpmnFileId) {
-    yield fetch(`/bpmn?id=${bpmnFileId}`, {
-      method: 'POST',
-      headers: {
-        'Content-Type': 'application/vnd.api+json',
-      },
-    });
-  }
-
-  @action
-  diagramUploaded(uploadedFileId) {
-    this.replaceModalOpened = false;
-    this.fetchLatestDiagramById.perform(uploadedFileId);
   }
 
   @action
@@ -295,52 +177,6 @@ export default class ProcessesProcessIndexController extends Controller {
 
     this.diagrams = undefined;
     this.latestDiagram = undefined;
-  }
-
-  // Tasks
-
-  @keepLatestTask
-  *fetchLatestDiagram() {
-    this.latestDiagramIsLoading = true;
-    this.latestDiagramHasErrored = false;
-
-    const query = {
-      reload: true,
-      page: {
-        number: 0,
-        size: 1,
-      },
-      'filter[processes][id]': this.model.process.id,
-      'filter[:or:][extension]': ['bpmn', 'vsdx'],
-      sort: '-created',
-    };
-
-    let diagrams;
-    try {
-      diagrams = yield this.store.query('file', query);
-    } catch {
-      this.latestDiagramHasErrored = true;
-    }
-    if (diagrams?.length) this.latestDiagram = diagrams[0];
-    else this.latestDiagramHasErrored = true;
-
-    this.latestDiagramIsLoading = false;
-  }
-
-  @keepLatestTask
-  *fetchLatestDiagramById(fileId) {
-    this.latestDiagramIsLoading = true;
-    this.latestDiagramHasErrored = false;
-
-    try {
-      this.latestDiagram = yield this.store.findRecord('file', fileId, {
-        reload: true,
-      });
-    } catch {
-      this.latestDiagramHasErrored = true;
-    }
-
-    this.latestDiagramIsLoading = false;
   }
 
   @keepLatestTask({
