+++ app/routes/processes/process/index.js
@@ -31,7 +31,6 @@ export default class ProcessesProcessIndexRoute extends Route {
   setupController(controller) {
     super.setupController(...arguments);
     controller.reset();
-    controller.fetchLatestDiagram.perform();
   }
 
   resetController(controller) {
